+++ cms_perf/setup/cli_parser.py
@@ -269,19 +269,6 @@ def compile_sensors(
     ]
 
 
-# CLI transformations
-@cli_call(name="max")
-def maximum(a: float, b: float, *others: float) -> float:
-    """The maximum value of all arguments"""
-    return max(a, b, *others)
-
-
-@cli_call(name="min")
-def minimum(a: float, b: float, *others: float):
-    """The minimum value of all arguments"""
-    return min(a, b, *others)
-
-
 if __name__ == "__main__":
     # provide debug information on the parser
     from ..sensors import sensor, net_load, xrd_load  # noqa  # pyright: ignore
